# Universe data frames normalization