--- a/Algorithm/QCAlgorithm.Python.cs
+++ b/Algorithm/QCAlgorithm.Python.cs
@@ -928,23 +928,20 @@ namespace QuantConnect.Algorithm
                 var requests = CreateBarCountHistoryRequests(new[] { universe.Symbol }, universe.DataType, periods, resolution, fillForward, extendedMarketHours,
                     dataMappingMode, dataNormalizationMode, contractDepthOffset);
                 // we pass in 'BaseDataCollection' type so the dataframe is properly created for the universe
-                return GetDataFrame(History(requests.Where(x => x != null)), symbolOnlyIndex: true, typeof(BaseDataCollection));
+                return GetDataFrame(History(requests.Where(x => x != null)), typeof(BaseDataCollection));
             }
             if (tickers.TryCreateType(out var type))
             {
                 var requests = CreateBarCountHistoryRequests(Securities.Keys, type, periods, resolution, fillForward, extendedMarketHours,
                     dataMappingMode, dataNormalizationMode, contractDepthOffset);
-                return GetDataFrame(History(requests.Where(x => x != null)),
-                    // TODO: Only universe? What about BaseDataCollection?
-                    symbolOnlyIndex: type.IsAssignableTo(typeof(Universe)),
-                    type);
+                return GetDataFrame(History(requests.Where(x => x != null)), type);
             }
 
             var symbols = tickers.ConvertToSymbolEnumerable().ToArray();
             var dataType = GetSymbolDataType(symbols, out var canonicalOption);
 
             var df = GetDataFrame(History(symbols, periods, resolution, fillForward, extendedMarketHours, dataMappingMode, dataNormalizationMode,
-                contractDepthOffset), symbolOnlyIndex: canonicalOption, dataType);
+                contractDepthOffset), dataType);
 
             if (canonicalOption)
             {
@@ -1001,16 +998,13 @@ namespace QuantConnect.Algorithm
                 var requests = CreateDateRangeHistoryRequests(new[] { universe.Symbol }, universe.DataType, start, end, resolution, fillForward, extendedMarketHours,
                     dataMappingMode, dataNormalizationMode, contractDepthOffset);
                 // we pass in 'BaseDataCollection' type so we clean up the dataframe if we can
-                return GetDataFrame(History(requests.Where(x => x != null)), symbolOnlyIndex: true, typeof(BaseDataCollection));
+                return GetDataFrame(History(requests.Where(x => x != null)), typeof(BaseDataCollection));
             }
             if (tickers.TryCreateType(out var type))
             {
                 var requests = CreateDateRangeHistoryRequests(Securities.Keys, type, start, end, resolution, fillForward, extendedMarketHours,
                     dataMappingMode, dataNormalizationMode, contractDepthOffset);
-                return GetDataFrame(History(requests.Where(x => x != null)),
-                    // TODO: Only universe? What about BaseDataCollection?
-                    symbolOnlyIndex: type.IsAssignableTo(typeof(Universe)),
-                    type);
+                return GetDataFrame(History(requests.Where(x => x != null)), type);
             }
 
             var symbols = tickers.ConvertToSymbolEnumerable().ToArray();
@@ -1018,7 +1012,7 @@ namespace QuantConnect.Algorithm
 
             // TODO: symbolOnlyIndex: canonicalOption? What about universe like FundamentalUniverse?
             var df = GetDataFrame(History(symbols, start, end, resolution, fillForward, extendedMarketHours, dataMappingMode,
-                dataNormalizationMode, contractDepthOffset), symbolOnlyIndex: canonicalOption, dataType);
+                dataNormalizationMode, contractDepthOffset), dataType);
 
             if (canonicalOption)
             {
@@ -1052,9 +1046,7 @@ namespace QuantConnect.Algorithm
             var requestedType = type.CreateType();
             var requests = CreateDateRangeHistoryRequests(symbols, requestedType, start, end, resolution, fillForward, extendedMarketHours,
                 dataMappingMode, dataNormalizationMode, contractDepthOffset);
-            return GetDataFrame(History(requests.Where(x => x != null)),
-                symbolOnlyIndex: symbols.Length > 0 && (IsCanonicalOption(symbols[0]) || requestedType.IsAssignableTo(typeof(Universe)) || (requestedType.IsAssignableTo(typeof(BaseDataCollection)) && symbols[0].SecurityType == SecurityType.Base)),
-                dataType: requestedType);
+            return GetDataFrame(History(requests.Where(x => x != null)), requestedType);
         }
 
         /// <summary>
@@ -1085,9 +1077,7 @@ namespace QuantConnect.Algorithm
             var requests = CreateBarCountHistoryRequests(symbols, requestedType, periods, resolution, fillForward, extendedMarketHours,
                 dataMappingMode, dataNormalizationMode, contractDepthOffset);
 
-            return GetDataFrame(History(requests.Where(x => x != null)),
-                symbolOnlyIndex: symbols.Length > 0 && (IsCanonicalOption(symbols[0]) || requestedType.IsAssignableTo(typeof(Universe)) || (requestedType.IsAssignableTo(typeof(BaseDataCollection)) && symbols[0].SecurityType == SecurityType.Base)),
-                dataType: requestedType);
+            return GetDataFrame(History(requests.Where(x => x != null)), requestedType);
         }
 
         /// <summary>
@@ -1165,9 +1155,7 @@ namespace QuantConnect.Algorithm
                     $"This could be due to the specified security not being of the requested type. Symbol: {symbol} Requested Type: {type.Name}");
             }
 
-            return GetDataFrame(History(requests),
-                symbolOnlyIndex: IsCanonicalOption(symbol) || type.IsAssignableTo(typeof(Universe)) || (type.IsAssignableTo(typeof(BaseDataCollection)) && symbol.SecurityType == SecurityType.Base),
-                dataType: type);
+            return GetDataFrame(History(requests), type);
         }
 
         /// <summary>
@@ -1786,18 +1774,18 @@ namespace QuantConnect.Algorithm
         /// <summary>
         /// Converts an enumerable of Slice into a Python Pandas dataframe
         /// </summary>
-        protected PyObject GetDataFrame(IEnumerable<Slice> data, bool symbolOnlyIndex, Type dataType = null)
+        protected PyObject GetDataFrame(IEnumerable<Slice> data, Type dataType = null)
         {
-            return PandasConverter.GetDataFrame(RemoveMemoizing(data), dataType, symbolOnlyIndex);
+            return PandasConverter.GetDataFrame(RemoveMemoizing(data), dataType);
         }
 
         /// <summary>
         /// Converts an enumerable of BaseData into a Python Pandas dataframe
         /// </summary>
-        protected PyObject GetDataFrame<T>(IEnumerable<T> data, bool symbolOnlyIndex)
+        protected PyObject GetDataFrame<T>(IEnumerable<T> data)
             where T : IBaseData
         {
-            return PandasConverter.GetDataFrame(RemoveMemoizing(data), symbolOnlyIndex);
+            return PandasConverter.GetDataFrame(RemoveMemoizing(data));
         }
 
         private IEnumerable<T> RemoveMemoizing<T>(IEnumerable<T> data)
@@ -1841,6 +1829,11 @@ namespace QuantConnect.Algorithm
             }
 
             using var _ = Py.GIL();
+
+            if (df.GetAttr("empty").GetAndDispose<bool>())
+            {
+                return df;
+            }
 
             using var renameArgs = new PyDict();
             using var canonicalName = "canonical".ToPython();