… Algorithm/QCAlgorithm.cs
… Algorithm/QuantConnect.Algorithm.csproj
… AlgorithmFactory/QuantConnect.AlgorithmFactory.csproj
… Common/Data/BaseData.cs
… Common/Data/Fundamental/Generated/FineFundamental.cs
… Common/Data/Market/DataDictionary.cs
… Common/Data/Market/Greeks.cs
… Common/Data/Market/OptionChain.cs
… Common/Data/Market/OptionChains.cs
… Common/Data/Market/OptionContract.cs
… Common/Data/Market/QuoteBar.cs
… Common/Data/Market/Tick.cs
… Common/Data/Market/TradeBar.cs
… Common/Data/UniverseSelection/BaseDataCollection.cs
… Common/Data/UniverseSelection/OptionUniverse.cs
… Common/ExtendedDictionary.cs
… Common/Python/PandasColumnAttribute.cs
… Common/Python/PandasConverter.DataFrameGenerator.cs
… Common/Python/PandasConverter.cs
… Common/Python/PandasData.DataTypeMember.cs
… Common/Python/PandasData.cs
… Common/Python/PandasIgnoreAttribute.cs
… Common/Python/PandasIgnoreMembersAttribute.cs
… Common/Python/PandasNonExpandableAttribute.cs
… Common/QuantConnect.csproj
… Common/Symbol.cs
… Engine/DataFeeds/BacktestingChainProvider.cs
… Engine/DataFeeds/DateChangeTimeKeeper.cs
… Engine/DataFeeds/SubscriptionDataReader.cs
… Engine/QuantConnect.Lean.Engine.csproj